Issue #80, add vote_count table to track daily changes in votes.
This table is needed to create the Threshold method. "prev_day_votes" variable will store the
number of votes the idea got the day before, and "curr_day_votes" will store the votes that
the idea will get in the new day. Each time the Threshold method is called, the "curr_day_votes"
column values will be set to 0, and the previous values will overwrite the column "prev_day_votes".